+++ ui/components/Confetti.tsx
@@ -1,4 +1,4 @@
-// @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'reac... Remove this comment to see the full error message
+import React from 'react'
 import { useState, useEffect } from 'react'
 import Confetti from 'react-confetti'
 import { useTimeout } from 'react-use'
@@ -13,6 +13,5 @@ export default function ConfettiComponent() {
     setHeight(window.innerHeight)
   }, [])
 
-  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
   return <Confetti width={width} height={height} recycle={!isComplete()} />
 }
