+++ ui/components/Layout.tsx
@@ -19,7 +19,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import Script from 'next/script'
-// @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'reac... Remove this comment to see the full error message
+import React from 'react'
 import { useState, useEffect } from 'react'
 // @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'reac... Remove this comment to see the full error message
 import Blockies from 'react-blockies'
@@ -27,18 +27,19 @@ import { useConnect, useEnsName, useDisconnect } from 'wagmi'
 import { nationToken } from '../lib/config'
 import { connectorIcons } from '../lib/connectors'
 import { useAccount } from '../lib/use-wagmi'
-// @ts-expect-error ts-migrate(2307) FIXME: Cannot find module '../public/logo.svg' or its cor... Remove this comment to see the full error message
 import Logo from '../public/logo.svg'
 import ErrorCard from './ErrorCard'
-// @ts-expect-error ts-migrate(6142) FIXME: Module './ErrorProvider' was resolved to '/Users/g... Remove this comment to see the full error message
 import { useErrorContext } from './ErrorProvider'
 import PreferredNetworkWrapper from './PreferredNetworkWrapper'
+
+type Indexable = {
+  [key: string]: any
+}
 
 const navigation = [
   {
     name: 'Start',
     href: '/',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <ViewGridIcon className="h-5 w-5" />,
   },
   /*{
@@ -49,44 +50,36 @@ const navigation = [
   {
     name: 'Claim airdrop',
     href: '/claim',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <SparklesIcon className="h-5 w-5" />,
   },
   {
     name: 'Lock tokens',
     href: '/lock',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <LockClosedIcon className="h-5 w-5" />,
   },
   {
     name: 'Liquidity rewards',
     href: '/liquidity',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <CurrencyDollarIcon className="h-5 w-5" />,
   },
   {
     name: 'Buy $NATION',
     href: `https://app.balancer.fi/#/trade/ether/${nationToken}`,
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <PlusIcon className="h-5 w-5" />,
   },
   {
     name: 'Homepage',
     href: 'https://nation3.org',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <HomeIcon className="h-5 w-5" />,
   },
   {
     name: 'Wiki',
     href: 'https://wiki.nation3.org',
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     icon: <NewspaperIcon className="h-5 w-5" />,
   },
 ]
 
-export default function Layout({
-  children
-}: any) {
+export default function Layout({ children }: any) {
   const router = useRouter()
   const { connectors, connect, error: connectError } = useConnect()
   //const { data: ensName } = useEnsName()
@@ -94,7 +87,7 @@ export default function Layout({
   const { disconnect } = useDisconnect()
   const [nav, setNav] = useState(navigation)
   const errorContext = useErrorContext()
-  const [account, setAccount] = useState('')
+  const [account, setAccount] = useState({} as any)
 
   // @ts-expect-error ts-migrate(2554) FIXME: Expected 1 arguments, but got 0.
   const { data: accountData } = useAccount()
@@ -103,300 +96,205 @@ export default function Layout({
   }, [accountData])
 
   return (
-    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
     <div className="mx-auto bg-n3bg font-display">
-      // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
       <Script src="https://cdn.splitbee.io/sb.js" />
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
       <div className="flex flex-col h-screen">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         <div className="navbar bg-base-100 border-slate-100 border-b-2 py-0 pl-0 lg:hidden">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <div className="navbar-start border-slate-100 pl-0">
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             <div className="w-80 border-slate-100 py-4 box-content">
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <div className="pl-6 pt-2 cursor-pointer">
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <div className="flex-none hidden lg:block">
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                   <Link href="/" passHref>
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                     <Image src={Logo} />
                   </Link>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </div>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <div className="flex-none lg:hidden">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <label
                     htmlFor="side-drawer"
                     className="btn btn-square btn-ghost"
                   >
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                     <MenuIcon className="h-8 w-8" />
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   </label>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </div>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </div>
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         </div>
 
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         <div className="drawer drawer-mobile w-full h-full grow max-h-screen flex-1">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <input id="side-drawer" type="checkbox" className="drawer-toggle" />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <div className="drawer-content flex flex-col overflow-auto">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
             <PreferredNetworkWrapper>{children}</PreferredNetworkWrapper>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <div className="drawer-side">
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             <label
               htmlFor="side-drawer"
               className="drawer-overlay z-10"
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             ></label>
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             <div className="bg-white w-80 flex flex-col justify-between pb-24 lg:pb-0 overflow-y-auto drop-shadow-md min-h-screen">
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <div className="mt-6 py-4 hidden lg:block">
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <div className="px-8 pt-2 cursor-pointer">
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                   <Link href="/" passHref>
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                     <Image src={Logo}></Image>
                   </Link>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </div>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </div>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <ul className="menu p-4 overflow-y-auto text-base-400 grow">
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                {nav.map((item: any) => <li
-                  className="mt-1 relative py-2"
-                  onClick={() =>
-                    // @ts-expect-error ts-migrate(2531) FIXME: Object is possibly 'null'.
-                    (document.getElementById('side-drawer').checked = false)
-                  }
-                  key={item.href}
-                >
-                  {item.href.charAt(0) === '/' ? (
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                    <Link href={item.href}>
-                      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                {nav.map((item: any) => (
+                  <li
+                    className="mt-1 relative py-2"
+                    onClick={() =>
+                      // @ts-expect-error ts-migrate(2531) FIXME: Object is possibly 'null'.
+                      (document.getElementById('side-drawer').checked = false)
+                    }
+                    key={item.href}
+                  >
+                    {item.href.charAt(0) === '/' ? (
+                      <Link href={item.href}>
+                        <a
+                          className={`py-4 ${
+                            router.pathname == item.href ? 'active' : ''
+                          }`}
+                        >
+                          {item.icon}
+                          {item.name}
+                          <ChevronRightIcon className="h-5 w-5 absolute right-4 opacity-50" />
+                        </a>
+                      </Link>
+                    ) : (
                       <a
                         className={`py-4 ${
                           router.pathname == item.href ? 'active' : ''
                         }`}
+                        href={item.href}
+                        target="_blank"
+                        rel="noopener noreferrer"
                       >
                         {item.icon}
                         {item.name}
-                        // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                        <ChevronRightIcon className="h-5 w-5 absolute right-4 opacity-50" />
-                      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                        <ExternalLinkIcon className="h-5 w-5 absolute right-4 opacity-50" />
                       </a>
-                    </Link>
-                  ) : (
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                    <a
-                      className={`py-4 ${
-                        router.pathname == item.href ? 'active' : ''
-                      }`}
-                      href={item.href}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {item.icon}
-                      {item.name}
-                      // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                      <ExternalLinkIcon className="h-5 w-5 absolute right-4 opacity-50" />
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                    </a>
-                  )}
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </li>)}
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                    )}
+                  </li>
+                ))}
               </ul>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <ul className="menu p-4 text-base-400">
                 {account?.address ? (
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <li>
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                     <label htmlFor="web3-modal">
-                      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                       <div className="mask mask-circle cursor-pointer">
-                        // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                         <Blockies seed={account?.address} size={12} />
-                      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                       </div>
                       {ensName
                         ? ensName
-                        : `${account.address.substring(
+                        : `${((account.address as string) ?? '').substring(
                             0,
                             6
                           )}...${account.address.slice(-4)}`}
-                      // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                       <ChevronDownIcon className="h-5 w-5 absolute right-4 opacity-50" />
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                     </label>
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   </li>
                 ) : (
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <li>
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                     <label
                       htmlFor="web3-modal"
                       className="btn btn-primary normal-case font-medium text-white modal-button"
                     >
                       Sign in
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                     </label>
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   </li>
                 )}
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </ul>
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
             </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
       </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
       <input type="checkbox" id="web3-modal" className="modal-toggle" />
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
       <label htmlFor="web3-modal" className="modal cursor-pointer">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         <label className="modal-box relative" htmlFor="">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <label
             htmlFor="web3-modal"
             className="btn btn-sm btn-circle btn-ghost absolute right-6 top-5"
           >
             ✕
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           </label>
 
           {account ? (
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
             <>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <h3 className="text-lg font-bold px-4">Account</h3>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
               <p className="p-4">Connected to {account.connector?.name}</p>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
               <ul className="menu bg-base-100 p-2 -m-2 rounded-box">
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <li key="address">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <a
                     href={`https://etherscan.io/address/${account.address}`}
                     rel="noreferrer noopener"
                     target="_blank"
                   >
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                     <UserIcon className="h-5 w-5" />
                     {ensName
                       ? ensName
-                      : `${account.address.substring(
+                      : `${((account.address as string) ?? '').substring(
                           0,
                           6
-                        )}...${account.address.slice(-4)}`}
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                        )}...${((account.address as string) ?? '').slice(-4)}`}
                   </a>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </li>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
                 <li key="logout">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  <a onClick={disconnect}>
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
+                  <a onClick={() => disconnect()}>
                     <LogoutIcon className="h-5 w-5" />
                     Log out
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   </a>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </li>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </ul>
             </>
           ) : (
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
             <>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               <h3 className="text-lg font-bold px-4">
                 Sign in by connecting your account
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </h3>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
               <p className="p-4">You can choose from these providers:</p>
               {connectError ? (
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <div className="alert alert-error mb-4">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <div>
-                    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                     <XCircleIcon className="h-5 w-5" />
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                     <span>{connectError?.message || 'Failed to connect'}</span>
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   </div>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </div>
               ) : (
                 ''
               )}
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
               <ul className="menu bg-base-100 p-2 -m-2 rounded-box">
                 {connectors.map((connector) => (
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                   <li key={connector.id}>
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                    <a
+                    <button
                       disabled={!connector.ready}
                       onClick={() => connect(connector)}
                     >
-                      // @ts-expect-error ts-migrate(7053) FIXME: Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-                      {connectorIcons[connector.name] ? (
-                        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                      {(connectorIcons as Indexable)[connector.name] ? (
                         <div className="h-5 w-5">
-                          // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                          <Image src={connectorIcons[connector.name]} />
-                        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                          <Image
+                            src={(connectorIcons as Indexable)[connector.name]}
+                          />
                         </div>
                       ) : (
-                        // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
                         <KeyIcon className="h-5 w-5" />
                       )}
                       {connector.name}
                       {!connector.ready && ' (unsupported)'}
-                    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                    </a>
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+                    </button>
                   </li>
                 ))}
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </ul>
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+
               <p className="px-4 mt-4">
                 New to Ethereum?{' '}
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <a
                   href="https://ethereum.org/wallets/"
                   rel="noreferrer noopener"
@@ -404,14 +302,10 @@ export default function Layout({
                   className="underline text-n3blue"
                 >
                   Learn more about wallets
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </a>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 .<br />
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <br />
                 By using this software, you agree to{' '}
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 <a
                   href="https://github.com/nation3/app/blob/master/LICENSE.md"
                   rel="noreferrer noopener"
@@ -419,32 +313,24 @@ export default function Layout({
                   className="underline text-n3blue"
                 >
                   its terms of use
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
                 </a>
                 .
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
               </p>
             </>
           )}
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         </label>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
       </label>
       {errorContext?.errors ? (
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         <div className="fixed md:right-8 md:bottom-8 md:left-auto bottom-0 left-0 right-0">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
           <div className="stack max-w-sm">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            {errorContext.errors.map((error: any) => <ErrorCard error={error} key={error.key} />)}
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+            {errorContext.errors.map((error: any) => (
+              <ErrorCard error={error} key={error.key} />
+            ))}
           </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         </div>
       ) : (
         ''
       )}
-    // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
     </div>
-  );
+  )
 }
