+++ ui/components/Quote.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Quote({ children, author }: any) {
+  return (
+    <div className="pl-3 border-l border-n3blue mb-4">
+      <p>{children}</p> <p className="text-n3blue text-sm mt-2">{author}</p>
+    </div>
+  )
+}
