+++ ui/package.json
@@ -9,10 +9,13 @@
   },
   "dependencies": {
     "@heroicons/react": "^1.0.5",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.34",
+    "@types/react-dom": "^18.0.4",
     "@uniswap/merkle-distributor": "^1.0.1",
     "daisyui": "^2.13.5",
     "ethers": "^5.6.0",
-    "next": "12.0.7",
+    "next": "12.1.5-canary.7",
     "react": "17.0.2",
     "react-animated-3d-card": "^1.0.2",
     "react-blockies": "^1.4.1",
@@ -25,6 +28,7 @@
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^3.2.0",
+    "@types/react": "^18.0.9",
     "autoprefixer": "^10.4.0",
     "eslint": "8.4.1",
     "eslint-config-next": "12.0.7",
