+++ ui/pages/liquidity.tsx
@@ -5,8 +5,8 @@ import {
   CalculatorIcon,
   InformationCircleIcon,
 } from '@heroicons/react/outline'
-// @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'reac... Remove this comment to see the full error message
 import { useState } from 'react'
+import React from 'react'
 import { useBalancerPool } from '../lib/balancer'
 import {
   balancerPoolId,
@@ -34,7 +34,7 @@ import Head from '../components/Head'
 import MainCard from '../components/MainCard'
 
 export default function Liquidity() {
-  // @ts-expect-error ts-migrate(2554) FIXME: Expected 1 arguments, but got 0.
+  // @ts-expect-error
   const { data: account } = useAccount()
 
   const { data: veNationBalance, isLoading: veNationBalanceLoading } =
@@ -42,8 +42,7 @@ export default function Liquidity() {
   const {
     poolValue,
     nationPrice,
-    // @ts-expect-error ts-migrate(2339) FIXME: Property 'loading' does not exist on type '{ poolV... Remove this comment to see the full error message
-    loading: poolLoading,
+    isLoading: poolLoading,
   } = useBalancerPool(balancerPoolId)
 
   const { data: poolTokenBalance, isLoading: poolTokenBalanceLoading } =
@@ -78,349 +77,272 @@ export default function Liquidity() {
   })
 
   const [depositValue, setDepositValue] = useState(0)
-  const [withdrawalValue, setWithdrawalValue] = useState(0)
+  const [withdrawalValue, setWithdrawalValue] = useState('0')
   const deposit = useDeposit(transformNumber(depositValue, 'bignumber', 18))
   const withdraw = useWithdraw(
     transformNumber(withdrawalValue, 'bignumber', 18)
   )
-  // @ts-expect-error ts-migrate(2554) FIXME: Expected 0 arguments, but got 1.
+
+  // @ts-expect-error
   const claimRewards = useClaimRewards(unclaimedRewards)
   const withdrawAndClaimRewards = useWithdrawAndClaim()
   const [activeTab, setActiveTab] = useState(0)
 
-  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-  return <>
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-    <Head title="$NATION liquidity rewards" />
+  return (
+    <>
+      <Head title="$NATION liquidity rewards" />
 
-    // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-    <MainCard title="$NATION liquidity rewards">
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      <p>
-        Provide liquitity in the pool and then deposit the pool token here.{' '}
-        // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-        <GradientLink
-          href="https://app.balancer.fi/#/pool/0x0bf37157d30dfe6f56757dcadff01aed83b08cd600020000000000000000019a"
-          text="Balancer pool"
-          textSize="md"
-        />
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <br />
-        Get up to {veNationRewardsMultiplier}x more rewards with $veNATION.{' '}
-        // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-        <GradientLink href="/lock" text="Get $veNATION" textSize="md" />
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      </p>
+      <MainCard title="$NATION liquidity rewards">
+        <p>
+          Provide liquitity in the pool and then deposit the pool token here.{' '}
+          <GradientLink
+            href="https://app.balancer.fi/#/pool/0x0bf37157d30dfe6f56757dcadff01aed83b08cd600020000000000000000019a"
+            text="Balancer pool"
+            textSize="md"
+          />
+          <br />
+          Get up to {veNationRewardsMultiplier}x more rewards with $veNATION.{' '}
+          <GradientLink href="/lock" text="Get $veNATION" textSize="md" />
+        </p>
 
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      <div className="stats stats-vertical lg:stats-horizontal shadow my-4">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="stat">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-figure">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <CurrencyDollarIcon className="h-8 w-8" />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-title">Total liquidity</div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-value">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <Balance
-              // @ts-expect-error ts-migrate(2362) FIXME: The left-hand side of an arithmetic operation must... Remove this comment to see the full error message
-              balance={transformNumber(poolValue, 'number', 0) / 1000000}
-              loading={poolLoading}
-              prefix="$"
-              suffix="M"
-              decimals={2}
-            />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
+        <div className="stats stats-vertical lg:stats-horizontal shadow my-4">
+          <div className="stat">
+            <div className="stat-figure">
+              <CurrencyDollarIcon className="h-8 w-8" />
+            </div>
 
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="stat">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-figure">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <TrendingUpIcon className="h-8 w-8" />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-title">Rewards APY</div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-value">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <Balance
-              balance={liquidityRewardsAPY}
-              loading={liquidityRewardsLoading}
-              suffix="%"
-              decimals={2}
-            />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      <div className="stats stats-vertical lg:stats-horizontal shadow mb-4">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="stat">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-figure text-primary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <SparklesIcon className="h-8 w-8" />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-title">Your veNATION</div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-value text-primary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <Balance
-              balance={veNationBalance}
-              loading={veNationBalanceLoading}
-              decimals={4}
-            />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="stat">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-figure text-secondary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <CalculatorIcon className="h-8 w-8" />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-title">Your boosted APY</div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-value text-secondary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <Balance
-              balance={boostedAPY || liquidityRewardsAPY}
-              suffix="%"
-              decimals={2}
-            />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      </div>
-      {canBoost && (
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="alert mb-4">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div>
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <InformationCircleIcon className="h-8 w-8 text-n3blue" />
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            <span>
-              You can boost your APY to{' '}
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-              <span className="text-n3blue font-semibold">
-                {transformNumber(
-                  // @ts-expect-error ts-migrate(2362) FIXME: The left-hand side of an arithmetic operation must... Remove this comment to see the full error message
-                  (transformNumber(liquidityRewardsAPY, 'number', 18) /
-                    10 ** 18) *
-                    potentialBoost,
-                  'string',
-                  2
-                )}
-                %
-              // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-              </span>
-              . To do so, claim your current rewards.
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            </span>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
-      )}
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      <div className="stats stats-vertical lg:stats-horizontal shadow mb-4">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="stat">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-figure text-secondary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <ActionButton
-              className="btn btn-primary normal-case font-medium grow"
-              action={claimRewards}
-            >
-              Claim
-            </ActionButton>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-title">Your rewards</div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-value text-primary">
-            // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-            <Balance balance={unclaimedRewards} decimals={4} />
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          </div>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="stat-desc">NATION tokens</div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      <div className="card bg-base-100 shadow">
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-        <div className="card-body">
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="tabs flex justify-center bg-white mb-4">
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            <a
-              className={`tab grow ${activeTab === 0 ? 'tab-active' : ''}`}
-              onClick={() => setActiveTab(0)}
-            >
-              Stake
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            </a>
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            <a
-              className={`tab grow ${activeTab === 1 ? 'tab-active' : ''}`}
-              onClick={() => setActiveTab(1)}
-            >
-              Unstake
-            // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-            </a>
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+            <div className="stat-title">Total liquidity</div>
+
+            <div className="stat-value">
+              <Balance
+                balance={
+                  (transformNumber(poolValue, 'number', 0) as number) / 1000000
+                }
+                loading={poolLoading}
+                prefix="$"
+                suffix="M"
+                decimals={2}
+              />
+            </div>
           </div>
 
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-          <div className="form-control">
-            {activeTab === 0 ? (
-              // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-              <>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <p className="mb-4">
-                  Available to deposit:{' '}
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                  <Balance
-                    balance={poolTokenBalance?.formatted}
-                    loading={poolTokenBalanceLoading}
-                  />{' '}
-                  LP tokens
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </p>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <div className="input-group">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  <input
-                    type="number"
-                    placeholder="Amount"
-                    className="input input-bordered w-full"
-                    value={depositValue}
-                    onChange={(e: any) => {
-                      setDepositValue(e.target.value)
-                    }}
-                  />
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  <button
-                    className="btn btn-outline"
-                    onClick={() =>
-                      poolTokenBalance &&
-                      setDepositValue(poolTokenBalance?.formatted)
-                    }
-                  >
-                    Max
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  </button>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </div>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <div className="card-actions mt-4">
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                  <ActionButton
-                    className="btn btn-primary normal-case font-medium w-full"
-                    action={deposit}
-                    approval={{
-                      token: balancerLPToken,
-                      spender: lpRewardsContract,
-                      amountNeeded: depositValue,
-                      approveText: 'Approve LP token',
-                    }}
-                  >
-                    Deposit
-                  </ActionButton>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </div>
-              </>
-            ) : (
-              // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-              <>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <p className="mb-4">
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                  Available to withdraw: <Balance balance={userDeposit} /> LP
-                  tokens
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </p>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <div className="input-group">
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  <input
-                    type="number"
-                    placeholder="Amount"
-                    className="input input-bordered w-full"
-                    value={withdrawalValue}
-                    onChange={(e: any) => {
-                      setWithdrawalValue(e.target.value)
-                    }}
-                  />
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  <button
-                    className="btn btn-outline"
-                    onClick={() =>
-                      userDeposit &&
-                      setWithdrawalValue(
-                        // @ts-expect-error ts-migrate(2554) FIXME: Expected 3 arguments, but got 2.
-                        transformNumber(userDeposit, 'string')
-                      )
-                    }
-                  >
-                    Max
-                  // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                  </button>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </div>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                <div className="card-actions mt-4">
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                  <ActionButton
-                    className="btn btn-primary normal-case font-medium w-full"
-                    action={withdraw}
-                  >
-                    Withdraw
-                  </ActionButton>
-                  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-                  <ActionButton
-                    className="btn btn-primary normal-case font-medium w-full"
-                    action={withdrawAndClaimRewards}
-                  >
-                    Withdraw all and claim
-                  </ActionButton>
-                // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-                </div>
-              </>
-            )}
-          // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
+          <div className="stat">
+            <div className="stat-figure">
+              <TrendingUpIcon className="h-8 w-8" />
+            </div>
+
+            <div className="stat-title">Rewards APY</div>
+
+            <div className="stat-value">
+              <Balance
+                balance={liquidityRewardsAPY}
+                loading={liquidityRewardsLoading}
+                suffix="%"
+                decimals={2}
+              />
+            </div>
           </div>
-        // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
         </div>
-      // @ts-expect-error ts-migrate(7026) FIXME: JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message
-      </div>
-    </MainCard>
-  </>;
+
+        <div className="stats stats-vertical lg:stats-horizontal shadow mb-4">
+          <div className="stat">
+            <div className="stat-figure text-primary">
+              <SparklesIcon className="h-8 w-8" />
+            </div>
+
+            <div className="stat-title">Your veNATION</div>
+
+            <div className="stat-value text-primary">
+              <Balance
+                balance={veNationBalance}
+                loading={veNationBalanceLoading}
+                decimals={4}
+              />
+            </div>
+          </div>
+
+          <div className="stat">
+            <div className="stat-figure text-secondary">
+              <CalculatorIcon className="h-8 w-8" />
+            </div>
+
+            <div className="stat-title">Your boosted APY</div>
+
+            <div className="stat-value text-secondary">
+              <Balance
+                balance={boostedAPY || liquidityRewardsAPY}
+                suffix="%"
+                decimals={2}
+              />
+            </div>
+          </div>
+        </div>
+        {canBoost && (
+          <div className="alert mb-4">
+            <div>
+              <InformationCircleIcon className="h-8 w-8 text-n3blue" />
+
+              <div>
+                You can boost your APY to{' '}
+                <div className="text-n3blue font-semibold">
+                  {transformNumber(
+                    ((transformNumber(
+                      liquidityRewardsAPY ?? 0,
+                      'number',
+                      18
+                    ) as number) /
+                      10 ** 18) *
+                      potentialBoost,
+                    'string',
+                    2
+                  ) + '%'}
+                </div>
+                . To do so, claim your current rewards.
+              </div>
+            </div>
+          </div>
+        )}
+
+        <div className="stats stats-vertical lg:stats-horizontal shadow mb-4">
+          <div className="stat">
+            <div className="stat-figure text-secondary">
+              <ActionButton
+                className="btn btn-primary normal-case font-medium grow"
+                action={claimRewards}
+              >
+                Claim
+              </ActionButton>
+            </div>
+
+            <div className="stat-title">Your rewards</div>
+
+            <div className="stat-value text-primary">
+              <Balance balance={unclaimedRewards} decimals={4} />
+            </div>
+
+            <div className="stat-desc">NATION tokens</div>
+          </div>
+        </div>
+
+        <div className="card bg-base-100 shadow">
+          <div className="card-body">
+            <div className="tabs flex justify-center bg-white mb-4">
+              <a
+                className={`tab grow ${activeTab === 0 ? 'tab-active' : ''}`}
+                onClick={() => setActiveTab(0)}
+              >
+                Stake
+              </a>
+
+              <a
+                className={`tab grow ${activeTab === 1 ? 'tab-active' : ''}`}
+                onClick={() => setActiveTab(1)}
+              >
+                Unstake
+              </a>
+            </div>
+
+            <div className="form-control">
+              {activeTab === 0 ? (
+                <>
+                  <p className="mb-4">
+                    Available to deposit:{' '}
+                    <Balance
+                      balance={poolTokenBalance?.formatted}
+                      loading={poolTokenBalanceLoading}
+                    />{' '}
+                    LP tokens
+                  </p>
+
+                  <div className="input-group">
+                    <input
+                      type="number"
+                      placeholder="Amount"
+                      className="input input-bordered w-full"
+                      value={depositValue}
+                      onChange={(e: any) => {
+                        setDepositValue(e.target.value)
+                      }}
+                    />
+
+                    <button
+                      className="btn btn-outline"
+                      onClick={() =>
+                        poolTokenBalance &&
+                        setDepositValue(poolTokenBalance?.formatted)
+                      }
+                    >
+                      Max
+                    </button>
+                  </div>
+
+                  <div className="card-actions mt-4">
+                    <ActionButton
+                      className="btn btn-primary normal-case font-medium w-full"
+                      action={deposit}
+                      approval={{
+                        token: balancerLPToken,
+                        spender: lpRewardsContract,
+                        amountNeeded: depositValue,
+                        approveText: 'Approve LP token',
+                      }}
+                    >
+                      Deposit
+                    </ActionButton>
+                  </div>
+                </>
+              ) : (
+                <>
+                  <p className="mb-4">
+                    Available to withdraw: <Balance balance={userDeposit} /> LP
+                    tokens
+                  </p>
+
+                  <div className="input-group">
+                    <input
+                      type="number"
+                      placeholder="Amount"
+                      className="input input-bordered w-full"
+                      value={withdrawalValue}
+                      onChange={(e: any) => {
+                        setWithdrawalValue(e.target.value)
+                      }}
+                    />
+
+                    <button
+                      className="btn btn-outline"
+                      onClick={() =>
+                        userDeposit &&
+                        setWithdrawalValue(
+                          transformNumber(userDeposit, 'string', 18) as string
+                        )
+                      }
+                    >
+                      Max
+                    </button>
+                  </div>
+
+                  <div className="card-actions mt-4">
+                    <ActionButton
+                      className="btn btn-primary normal-case font-medium w-full"
+                      action={withdraw}
+                    >
+                      Withdraw
+                    </ActionButton>
+
+                    <ActionButton
+                      className="btn btn-primary normal-case font-medium w-full"
+                      action={withdrawAndClaimRewards}
+                    >
+                      Withdraw all and claim
+                    </ActionButton>
+                  </div>
+                </>
+              )}
+            </div>
+          </div>
+        </div>
+      </MainCard>
+    </>
+  )
 }
